login+crud

+++ app/Http/Controllers/UserController.php
@@ -6,14 +6,25 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 
+//use Request;
+
 use App\User;
 
 use Exception;
 
-class UserController extends Controller
-{
+use \Firebase\JWT\JWT;
 
+class UserController extends base
+{
+    private $key = 'danj71';
+
+   
     private $path = 'user';
+
+    public $tokenver = '';
+
+
+
     /**
      * Display a listing of the resource.
      *
@@ -23,6 +34,7 @@ class UserController extends Controller
     {
         //
         $data = User::all();
+
         //return $data;
         return view('index', compact('data'));
     }
@@ -75,9 +87,22 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        $data = User::find($id);
+        // var_dump($this->tokenver);
+        // $head = apache_request_headers();
+        // $head['Authorization'] = $this->tokenver;
+        if($this->check()){
+            //var_dump($this->idToken);
 
-        return $data;
+            if ($this->idToken == $id){
+
+                $data = User::find($id);
+                echo $this->tokenver;
+
+                return $data;
+            }else 
+                return('acceso denegado');
+        }else 
+            return ('autentificacion fallida');
     }
 
     /**
@@ -88,8 +113,12 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $data = User::find($id);
-        return view('edit', compact('data'));
+        if($this->check()){
+            if($this->idToken == $id){
+                $data = User::find($id);
+                return view('edit', compact('data'));
+            }
+        }
     }
 
     /**
@@ -101,26 +130,35 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $user = User::find($id);
-        var_dump($request->name);
-        var_dump($request->email);
+        if($this->check()){
+
+            if ($this->idToken == $id){
+
+                $user = User::find($id);
+                var_dump($request->name);
+                var_dump($request->email);
+                var_dump($request->header('key'));
+                    
             
-    
-         if($request->name != null){
-            $user->name = $request->name;
-         }
-         if($request->email != null){
-            $user->email = $request->email;
-         }
-         if($request->password != null){
-            $user->password = $request->password;
-         }
-         if($request->avatar != null){
-            $user->avatar = $request->avatar;
-         }
-       
-         $user->save();
-         return redirect()->route('user.index');
+                if($request->name != null){
+
+                    $user->name = $request->name;
+                }
+                if($request->email != null){
+                    $user->email = $request->email;
+                }
+                if($request->password != null){
+                    $user->password = $request->password;
+                }
+                if($request->avatar != null){
+                    $user->avatar = $request->avatar;
+                }
+               
+                $user->save();
+                return redirect()->route('user.index');
+            }
+        }
+        
     
             
         
@@ -142,14 +180,17 @@ class UserController extends Controller
         // catch(exception $e){
         //     return ('Usuario no encontrado');
         // }
+        if ($this->check()){
+            if ($this->idToken == $id){
+                $data = User::find($id);
 
-        $data = User::find($id);
-
-        if ($data){
-            $del = User::destroy($id);
-            return('Usuario borrado');
-        }else{
-            return('Usuario no encontrado');
+                if ($data){
+                    $del = User::destroy($id);
+                    return('Usuario borrado');
+                }else{
+                    return('Usuario no encontrado');
+                }
+            }
         }
 
 
@@ -158,6 +199,62 @@ class UserController extends Controller
 
     public function login (Request $request)
     {
-        return ("login hecho");
+        $tokentohead = apache_request_headers();
+
+        $email = $request->email;
+        $password = $request->password;
+
+        $user = User::where('email', $email)->get();
+        //var_dump($user);
+        
+        
+        if(isset ($email)){
+
+            if (isset($password)){
+
+                if(!empty($user)){
+
+                    foreach ($user as $key => $value)
+                    {
+                        $id = $user[$key]->id;
+                        $email1 = $user[$key]->email;
+                        $pass = $user[$key]->password;
+                        // var_dump($id);
+                        //exit;
+                    }
+                    // var_dump($id);
+
+
+                }else
+                    return ('Usuario no encontrado');
+
+                
+
+                if ($email == $email1 && $password == $pass){
+
+                    $token = array(
+                        "id" => $id, 
+                        "email" => $email1, 
+                        "password" => $pass
+                    );
+                    
+                    $jwt = JWT::encode($token, $this->key);
+
+                    $tokentohead['Authorization'] = $jwt;
+                    $this->tokenver = $jwt;
+                    var_dump($this->tokenver);
+
+
+                    // return $this->notice($code = 'SUCCESSFUL ACTION', $message = $jwt);
+                    return array('code' => 200, 'token' => $jwt, 'id' => $id);
+                }else
+                    return('Contraseña erronea');
+
+            }else 
+                return('Falta rellenar contraseña');
+
+        }else
+            return('Falta rellenar email');
+        
     }
 }
